lamatic_bridge/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
loguru==0.7.2
qdrant-client==1.7.0
